backend/requirements.txt
django==5.2.5 
djangorestframework==3.14.0 
Pillow==10.0.0 
WeasyPrint==60.0 
django-cors-headers==4.3.1 
qrcode==7.4.2 
